Extract shared worker venv cache support
